requirements.txt
Django==3.2
psycopg2-binary>=2.8
djangorestframework==3.12.4
